+++ src/main/java/app/utils/processor/impl/user/UserDeleteProcessor.java
@@ -4,9 +4,12 @@ import app.service.UserService;
 import app.service.impl.UserServiceImpl;
 import app.utils.SecurityContextHolder;
 import app.utils.processor.Processor;
+import lombok.extern.log4j.Log4j2;
 
 import static app.entity.user.enums.Role.ADMIN;
+import static app.utils.SecurityContextHolder.isCurrentUserSignedIn;
 
+@Log4j2
 public class UserDeleteProcessor implements Processor {
     private final UserService userService = new UserServiceImpl();
 
@@ -17,7 +20,7 @@ public class UserDeleteProcessor implements Processor {
 
     @Override
     public void process(String command) {
-        if (isAdmin()) {
+        if (isCurrentUserSignedIn() && isAdmin()) {
             final Long id = Long.parseLong(command.split(" ")[1]);
 
             userService.deleteById(id);
